Update to v2.1.1 of https://github.com/nlohmann/json (#386)
* Update to v2.1.1 of https://github.com/nlohmann/json

The old version we were using has pathalogical performance issues when
parsing numbers in large inputs. This was leading to JSON parsing taking
tens of minutes in the data connector. The same inputs now take just a
handful of seconds.

* base64-encode binary data before serializing it to JSON

JSON is limited to UTF-8, so attempting to store arbitrary binary data
in it is asking for trouble. It sometimes results in JSON data being
written that cannot be parsed. Encoding the data to base64 avoids this.