Don't reuse list notifiers which had all callbacks removed
If a transaction is processed while a notifier has no callbacks, no work is
done for that notifier and it gets into an inconsistent state. Adding a new
callback to that notifier thus results in incorrect results from the first
notification sent due to that the cached state is not properly reinitialized.
Work around this for now by just not reusing list notifiers.

This problem does not apply to Results notifiers due to that the changeset in
the Initial notification for those is not meaningful.