Rework the global notifier
Shift around some responsibilities between AdminRealmLister and GlobalNotifier
so that Adaptor only depends on the former rather than both of them, which
removes the need for GlobalNotifier to specifically support Adaptor.

Remove the changeset calculation worker queue. This doesn't work with the
future design of performing notifications in multiple processes, and turned out
to be very low value (even with only a single notification thread the new code
should be overall higher throughput than before).

Eliminate the use of `enable_shared_from_this()` and make GlobalNotifier and
Adaptor deal with keeping the things needed for callbacks alive rather than
bleeding that into the public interface.

Return a promise from Sync.addListener() which is resolved when the admin realm
has been downloaded and listening has begun, or is rejected with an error if
that fails. Previously errors were being silently discarded, and the tests had
race conditions where the listener was being initalized after the test which
created it completed.