Validate that all local transactions have been processed for upload progress notifications
When adding a non-streaming upload progress notifier immediately after
performing a local write, it's possible to receive a progress notification
which does not include that transaction in the uploadable bytes. Handle this by
ignoring notifications which have a snapshot version less than that of the last
local transaction made at the time the notifier was created.