Simplify ref->ptr translation and make the common case inlinable
Translation can now be fully inlined, even when encryption is
compiled in.

Sharing of mappings have been removed in preparation for threading work.

File growth is now much closer to what is actually needed
(no longer a requirement that the file is at least 8x max allocation)

Not fully functional on windows yet.

Change-Id: I3e891f1f1c33ad8114216da2bed008e169f255d5