Avoid using RealmCoordinator on the worker thread.
Using RealmCoordinator on the background thread can result in the
thread holding the last remaining reference to the coordinator. This
results in the coordinator being deallocated on the work queue thread,
while the work queue is still running, leading to a crash.

We instead use lower-level core APIs for the schema changes and data
access we perform.