Fix things with core v5.2.0.
The parser now returns an ordering in addition to the query, so handle that.

Serializing unsupported queries now throws an exception. We allow this
to propagate from `partial_sync::subscribe` as we consider it to be
programmer error, rather than a runtime error that would otherwise be
reported via the notification callback