Replace Table::optimize with Table::enumerate_string_column
The reason why the optimize functionality is not yet taken into use is
that the concept is not clear. It is not clear when optimize should be
performed, and we are not sure that the columns that are converted into
string enums, are the columns that will benefit in the longer run. We
think it is better that the user decides which columns to internalize
based on the use case. Only columns that are designed to hold a smaller
set of strings should be optimized.

Change-Id: I90773ba1cdef4a359c1fdf70e7e9ee75f3ae171f