Fix creating encrypted files on macOS
The allocation size was not adjusted for the additional pages needed for
encryption headers, resulting in the call to `m_file.prealloc(initial_size)`
during file initialization truncating away the previously-written data.
This corrects that oversight, and also makes `File::prealloc()` a no-op if the
size is smaller than the file's current size, to match the `posix_fallocate()`
behavior.