Fix partial sync schema initialization
`initialize_schema` assumed it would always be responsible for creating
the `__ResultSets` table. With sync v3.0.0-beta.10, the server now
eagerly creates the `__ResultSets` table when a partially-synced Realm
is opened, so that assumption is no longer valid. If a partially-synced
Realm is opened with `asyncOpen`, the server will have created the
schema and `initialize_schema` will still need to add the `name` column
to `__ResultSets`.