Add a 'delete' event to the global notifier
This event reports when a Realm matching the regex has been deleted from the
server. It is delivered only after all pending changes for the Realm have been
processed, and the local copy of the Realm is deleted from disk after the
handler completes.

The design of the GN has been adjusted a bit to track all realms seen even if
there is no change listener for them, but still open and sync realms which have
a change listener. In the process this sorts out some edge-cases around realm
deletion that I suspect did not work correctly.