Update permissions tests to work with sync 3.8.8
The server now tracks admin privileges on a partial-view basis rather than
based on the privileges of the oldest still-connected session, which makes
accessing a partial realm with a single user being both an admin and non-admin
no longer works.

Work around this by just using separate admin and non-admin users in the tests.