std::hash implementation for StringData (#3132)
Fixes #2929

I have copied what libc++ does for std::string. On 64-bit platforms, CityHash64 is used. On 32-bit platforms, Murmur2 is used.

CityHash is competitive with very fast hash functions, such as XXHash.

Note: This can all go away once we switch to C++17, where the standard library is required to provide an implementation of std::hash for std::string_view.

The problem this solves is that we can't currently use StringData as the key type in std::unordered_map, but have to always use std::string, which is suboptimal because it needs to own its data.

To reviewers: Please do not feel obligated to review the details in the implementation of CityHash64. It's pretty long. ;-) It should be a straight up copy of the implementation in libc++ (with adjustments for code style).