Encryption page reclamation (#3137)
This PR adds a page reclaim daemon thread, which will periodically reclaim decrypted pages and release them.

The process is controlled by a governor function which (once registered) is called periodically to provide a target for how many encrypted pages are kept. The reclaimer thread then works to meet that target.

The reclaimer is limited in the sense, that it cannot reclaim any page which has been visited during a still active transaction. This means --- for example --- that a transaction which scans the entire database will still need to have the full decrypted database in memory, since reclamation only starts after completion of the transaction.