Disable change notifications for SyncMetadataManager
The libuv implementation of EventLoopSignal always invokes notifications on the
main thread rather than the thread where the Realm was created. This works fine
for user-created Realms since we don't support using multiple threads at once
in realm-js, but it means that when we create our own Realms on background
threads we need to ensure that they never have an autorefresh notification
fire or we'll hit an incorrect thread error.

SyncMetadataManager's Realm is always short-lived and typically do not get any
notifications simply because it was closed before they could happen, but a
client reset error will sometimes trigger one. We don't want the notifications
anyway, so just disable them.