Optimise find_first on small string arrays (#3200)
Optimise find_first on small string arrays

This method showed up in a performance test (sync's bench-partial-sync). It's not necessarily super
slow, but since it's a very commonly used function, every optimisation counts.

The idea with using memcmp is that the compiler has the option to use a faster implementation if it
is available but in the worst case, I don't think it would be slower than our previous implementation
of character by character comparison.

In the test I was running (mac, release mode) it saves about 100 ms off 6.66 total time: ~3%
improvement.