Optimise queries search for a chain of OR strings (#3250)
This is a performance enhancement motivated by users who are generating queries with many string
comparisons on a single column, for example from cocoa's "IN" queries.

The idea is to combine string equality conditions from a single "OR" query node and store them in an
unordered_set. With N elements to search, and C conditions, the runtime changes from O(N*C) to
O(N). The added benchmark goes from 30 seconds to 2 seconds. This change does not try to optimise 
indexed columns which should be running O(log(N)*C). The benchmark with indexes turned on runs in 
3.5 seconds. Since N is likely the dominant term, using indexes should still be fastest in practice when
compared to this optimisation.