Fix unsubscribing from a subscription which hasn't been created yet
Unsubscribing from a Subscription which was still in the Creating state only
worked correctly if the caller held onto a reference to the Subscription object
until it made it to the Invalidated state (as the existing tests happen to). If
it was destroyed we would either fail to remove the subscription or even
outright crash depending on when the Subscription object was deallocated.