Add support for backlink inclusions (#3261)
The new IncludeDescriptor class gives a way to communicate the backlink paths that should be reported, and it is implemented in a similar way to sort/distinct/limit.

For integration into the js binding, a parse_include_path function is provided so that include paths can be added via a separate subscription builder method. The intention is for js to throw if a normal query contains an include descriptor and only add the included clause to queries that are sent to the server by a subscription. However, core does not throw by default, so it is up to the js binding to decide if "INCLUDE" should be allowed to pass through in normal queries.
* Beautification