Fix an incorrect debug mode assertion (#3276)
Table comparison is the wrong thing to do here because the implementation of Table::operator== checks for data equality across all rows. What was meant is that this is the same table, regardless of the data contained in it. A more lightweight check is now used just verifying name equality.