Fix unbound growth of freelist (#3278)
Problem with previous algorithm was that it produced more small blocks
than could afterwards be consumed. So we would have an ever increasing
number of small blocks.

Instead of using the first block that is just big enough the algorithm
is changed so that we will accept either a perfect match or a block that
has at least twice the size as the size that is requested.