Fix a race condition in some tests
If the Realm connected to the server fast enough it could download the
server-created subscriptions before the local ones were created, both resulting
in subscriptions being in the wrong index and the column order being wrong. Fix
this by always downloading the server subscriptions first and making the tests
handle that.