Avoid calling get_shared_realm() on background threads
The libuv implementation of EventLoopSignal only works when it's constructed on
the thread which the default loop is running on, as only uv_async_send() is
thread-safe. This isn't a problem for normal usage because node is
single-threaded, but it means that creating a Realm instance ourselves on a
background thread (such as when the sync worker thread calls a callback) does
bad things and we need to not do that.