Defer binding the async opened Realm to an execution context
When we open the Realm on the background thread, rather than bind it to that
thread, leave it in an unbound state and pass it to the callback as a
ThreadSafeReference<Realm>, which the SDK can then bind to a thread at a later
time.

This works around an issue in the libuv version of EventLoopSignal where
creating it on a background thread doesn't work, and makes async open work
better with Realm caching without requiring that the SDK reopen the Realm on
the destination thread, which is a potentially somewhat expensive operation.