Add workaround for Sync issue
Sync version 4.7.7 and earlier has the problem that the instruction 'ArrayInsert'
only inserts an empty row in the list and ignored the value provided.

This is not a problem with Core-5 as this instruction never carries a value.

The workaround splits the insert command in a "insert null" and a "set".
This is done only for string values.

Change-Id: I1ed15107fde3f9f6ef956ca480a7809ac1676ac2