Properly propagate update_only_diff through List::assign()
Calling Object::create(update_only_diff=true) on an object type which had a
List of objects would fail to propagate update_only_diff=true to the sub-calls
of Object::create(), resulting it those objects not being diffed and incorrect
change notifications being generated.

Because passing around 3 booleans all over the place is incredibly awkward and
error-prone, I turned create/update/update_only_diff into a single enum that
represents all of the valid combinations of those. This should make it
impossible to have a similar bug elsewhere, as it'd have to fail to propagate
the policy entirely and that's much more likely to be caught by tests.