"Fix" sorting float columns containing NaN
Not handling NaN in the sort comparator results in it violating one of
std::sort's preconditions (that the data have a total ordering), which with
libc++'s implementation can result in out-of-bounds reads.

Fix this by generalizing the behavior for nulls (which we store as a tagged
NaN) to all NaNs and using int comparisons to sort so that the NaNs are
ordered.