Fix changing primary keys in existing tables
Don't add indexes to string columns when setting them as the primary key (and
remove any existing index) to match the behavior of creating new tables with
string PKs.

Don't use ObjectIDs to validate PK uniqueness for non-indexed columns as
changing the PK makes it so that the PK and object keys are no longer
connected. This probably isn't otherwise an issue because ObjectIDs are only
used for synced realms, and the PK can't be changed on synced tables anyway.