Fix incorrect query results for NOT on a LinkList
NOT queries cache information about what range of rows has already been
checked, and when checking a range of rows immediately before the cached range
which does not contain any matches, find_first() would return the first row
within the cached range which matched. Since this wasn't not_found, this was
considered a match even though it wasn't in the range [start, end).