Add locking to frozen Results
Frozen Results aren't thread-confined, so we need to guard access to all of the
member variables which can be assigned to after construction. Locking adds a
measurable amount of overhead (~!0% on the applicable Cocoa performance tests),
so it's only actually done for frozen Results and the lock is a no-op
otherwise.

This introduces the use of clang's thread safety annotations to Results and a
wrapper type for std::mutex with the applicable annotations, which makes clang
warn at compile time is a variable guarded by a mutex is accessed without
acquiring the mutex and if a mutex is acquired in a function which is called
with that mutex already held.