Fix a thread sanitizer false positive
TSan complains about reading the elements of the m_ref_translation_ptr array
being a data race because the write to the element in the array could be
reordered to after the write to m_ref_translation_ptr. This is a false-positive
as release-acquire ordering on m_ref_translation_ptr prevents this.

To fix this, add explicit annotations to help tsan out. Only reading
m_ref_translation_ptr once inside Allocator::translate() is needed for this,
but also is a trivial optimization (each access of m_ref_translation_ptr is a
memory read while accessing the non-atomic local isn't).