Restore support for notifications on sorted primitive lists
As Lists can no longer be accessed via Query/TableView, Lists which have been
wrapped in a Results (normally for sort/distinct) need their own special
handling for notifications.

The logic for this is very similar to sorted notifications in core 5, with
manual updating of row indices to reflect insertions/deletions prior to
diffing.