Fix some bugs with creating ThreadSafeReference inside write transactions
With core 5 this was forbidden, but with core 6 it mostly works fine. The
exception is with objects which didn't exist yet at the start of the write
transaction, as the object key doesn't actually exist in the stored version.
For `ThreadSafeReference<Object>` and `ThreadSafeReference<List>` this has a
straightforward fix of refreshing the target Realm if it's at the version which
the relevant write was started at before trying to hand over the object/list.
For Results derived from a List or LinkingObjects, Query doesn't expose the
required members to do this so just disallow it for now.