Simplify query node cloning
Cloning a Query involved passing around a Transaction to the tree of query
nodes that was never actually used for anything. This was a remnant of the core
5 logic where we held onto references to Columns which had to be updated when
performing handover, but query nodes no longer hold onto any persistent
references that tie them to a specific Transaction.

There's one semi-exception: subqueries. Fortunately, we already don't actually
support subqueries which would require handing over any members in the child
query. A subquery derived from a LnkLst or TableView isn't really a meaningful
concept, can't be expressed in the query language, and would have given
incorrect results if assembled manually.