Rework thread checking and event loop signaling
Combine these two into a "Scheduler" interface which covers both concepts to
support Realms which are confined to a scheduler which isn't an event loop
running on a single thread. This also subsumes some of the special-casing for
frozen Realms by making them just use a no-op scheduler that doesn't perform
thread checking or support notifications.

Add a dispatch-queue based scheduler on Apple platforms which posts
notifications to a GCD queue rather than a run loop and confines the Realm to
that queue (which may not always be on the same thread).