Upgrade non-synced realms with string primary keys correctly (#3786)
Non-syncked realm does not have an !OID column with the key to be used, so
objects were just created with sequential object key. This breaks the
invariant that objects with a string primary key should have their object
key generated from the pk value.

This caused the handling of links to be changed. Now we store the original
row index in a temporary column so that it can be used to find the right
object to link to. We can simply search for the value stored in the old
link column.