Report whether create_object_with_primary_key() created a new object
The object store code which calls this needs to know if it's populating a new
object or modifying an existing one. Currently it does that by looking up the
object first and only calling create_object_with_primary_key() if it doesn't
exist, which results in some duplicated code and a redundant lookup.