Fix handling of primary keys in migrations
Creating objects inside migrations which change the primary key property of the
object type being created didn't work properly because we called
Table::create_object_with_primary_key(), and the Table doesn't actually have a
primary key until after the migration ends.

Modifying primary key values for existing objects inside migrations almost
entirely worked, but in a very fragile way. In particular, for string primary
keys without indexes (which should be all of them, but the core 5 -> core 6
migration unintentionally left existing indexes in place) it only worked
because we corrupted the table's state in a way that resulted in the check to
see if a primary key was already in use giving false positives, which resulted
in using a code path that accidentally worked.

To fix this, trying to mutate a primary key inside a migration will now remove
the primary key from the table and restore it at the end of the migration,
making things behave the same as when adding a primary key to an existing
table.