Fix reattaching the allocator after compact
SlabAlloc::detach() sets m_youngest_live_version to 0 and attach() doesn't
touch it (as it comes from the lockfile rather than anything it has access to),
so compact needs to set m_youngest_live_version after attaching the allocator
to the new Realm file. Failing to do this resulted in the first ref translation
table being deallocated after the first write transaction even if it was still
in use, resulting in a use-after-free for whatever was using it.