Fix a version pinning issue with Results
Core 6 changed how TableViews were passed from ResultsNotifier to Results to be
pull-based rather than push based. With Core 5, ResultsNotifier mutated the
parent Results to update the TableView as part of change notification delivery,
while now Results gets the TableView from the ResultsNotifier when the user
reads from the Results and forces a re-evaluation of the query. This change
simplified the data flow a lot, but meant that if the Results is never actually
read from, the TableView would sit around in the ResultsNotifier forever, and
continue to pin that version.

To fix this, add a check to see if there's a stale TV that went unused and if
so discard it and release the version pin.