Don't try to deliver async query results inside a write transaction
When we import a TableView into an active write transaction we don't actually
know if the TV is up to date, as we don't keep track of which tables have been
modified so far in the transaction. This means that we just can't use the
background TV if it isn't picked up before starting the write transaction, or
it may have stale results if it's imported after a change that effects it.