Fix deadlocks on exFAT
exFAT doesn't actually allocate an inode for a file until it's non-empty, which
results in our checks based on that to fail. The most immediate implication of
this was that all interprocess mutexes ended up sharing a single local mutex
and deadlocking if more than one is ever locked at once.

Fix this by just ensuring that files are non-empty.