Skip validating dispatch queue types on older OSes
There's no public API to find out if a dispatch queue is serial or concurrent.
On more recent OS versions we can check the class to determine it, but this
doesn't work on older versions.

It seems fairly safe to skip the checks on old OSes since the only way it'd go
uncaught is if a developer tests exclusively on very old devices (and never in
the simulator or recent devices) or has separate code paths for old OS
versions.