Don't use search indexes for non-batchable queries
Inside a Not() or when a query is limited to a view we evaluate a single row at
a time. This makes it so that queries are always O(N*M) and we can't reduce the
asymptotic bound using a search index. Merging OR groups reduces M (potentially
all the way to 1), so in this case it's worth doing even though it means not
using indexes.