Correctly update queries on indexed string columns going from multiple matches to one
When one run of a query on an indexed string column matched multiple objects
and then a subsequent one matched one, m_index_matches was not set back to null
and so would incorrectly continue to be used in index_based_aggregate().

This would result in either incorrect results (if the first matching object was
mutated to no longer match) or throwing "key not found" if the first matching
object was deleted.