Set the supported platforms for the Swift package (#4039)
Xcode's Archive command has started building packages for all platforms they claim to support even if the app doesn't consume those platforms. In our case, this results in errors when building for 32-bit iOS because that platform does not fully support C++17. Fix this by setting the deployment target for iOS to 11, which dropped support for 32-bit.