Add missing return in empty-line case of watch() decoder
This was actually mostly harmless because real consumers wouldn't actually
pass in a null string_view, or one where reading from line[0] would trap, and
the later lines of the function are no-ops for an empty line. However, the
read of line[0] triggered an assert in a debug build with MS STL string_view,
which is how this was found.