Fix a race condition in notifier initialization (#4234)
Newly created notifiers copied their Query from the source Realm on the worker
thread without locking the source Realm, which could race with things that
invalidated the source Realm.

To fix this, switch to copying the required data to a temporary Transaction
owned by the notifier while still on the source thread. This increases the
amount of work done on the source thread to create a notifier (as it has to
create a new Transaction each time), but eliminates the need to hold the lock
while parsing the transaction logs for new notifiers, so it may turn out to be
a net reduction in main-thread blockage.