Fix invalidation of objects with embedded objects (#4240)
The issue was an early out on the first link column. It should be an early out on the first link column linking to an embedded table. So if the first column was not linking to an embedded table, the function would falsely report that the table does not have embedded objects.