Fix ObjKey generation after file format upgrade (#4296)
For some yet unknown reason, the stored sequence_number could have the value
0x400000000000000. This would break the assertion in GlobalKey::get_local_key().

The implementation is now changed so that the generated sequence number is only
32 bit. So even if we find this stray number, the assertion will not be broken.

We also ensure that the sequence number is not updated if the table has a primary
key column. If we set the sequence number to be one bigger than the biggest
ColKey, then it will in many cases be out of range.