Fix incorrect reads from Results backed by a Table
Table::Iterator::operator[] indexed relative to the current iterator position,
but Results (the only thing using it) expects it to be an absolute position.
Results only ever created begin() iterators and did not increment them so this
usually happened to work, but creating a new object which was inserted into the
Table at position zero would result in the iterator having a position of 1, and
all subsequent uses of the iterator would give incorrect results.

As the previous semantics were correct for operator[] and just didn't match
what Results actually needs, this replaces operator[] with calling `go()` and
making `go()` move the iterator to the given absolute position.

operator[] also failed to actually provide the optimization which the iterator
was being used for; m_leaf_invalid was always set to to true and so it reloaded
the leaf on every access. Fixing this cuts the runtime of
IterateTableByIteratorIndex from 100ms to 2ms on my machine.