Update all involved allocators before performing queries over links
This is an attempt at a more general solution to the same problem as f045da0ac.
There are several places where queries are using the allocator from one Table
while accessing another Table. Rather than attempting to validate that all of
them are correct, instead update each Table's WrappedAllocator before running
the query so that it doesn't matter which one is used.