Fix more version skipping logic errors
There were two bugs here:

1. The logic for whether we block until notifiers are ready when beginning a
write transaction and whether we set `m_notifier_skip_version` in
`commit_write()` (which depends on us having blocked when starting the write)
were mismatched. We don't block for notifiers without callbacks as the only
purpose of those notifiers is to rerun the query on a background thread to
avoid spending time on the main thread, so waiting would be counterproductive.
This means that we also need to not set the skip version if we only have
notifiers without callbacks. This probably did not cause any functional
problems, but violated some of the invariants of notification skipping and so
assertions checking those invariants failed.

2. The latest snapshot version needs to be acquired while the notifier lock is
held to ensure that we don't get mismatched values for it and the skip version.
368f0c7e5 fixed this in one of the code paths, but another code path implicitly
obtained the latest snapshot version via `advance_read(VersionID{})` without
holding the lock. The fix for this is to just always obtain the "final" version
before releasing the lock, which ends up simplifying the code a little as well.