Fix the check for if ResultsNotifier needs to rerun queries
The previous check would only skip rerunning the query (and sort etc.) if both
none of the tables in the query were modified and if the query used every table
reachable from the source table, which in many cases meant that it simply reran
after every commit. Instead we want to only rerun if any of the tables used by
the query were modified, and only check the objects from the previous run for
modification otherwise.