Consolidate the code for Collection, LinkList and LinkSet in Results
In addition to eliminating some duplicated code and mostly simplifying things,
this also fixes a pile of bugs with Results wrapping a Dictionary of links and
makes the API less error-prone. It was possible to construct a Results from a
CollectionBase with a type of Object (as Dictionaries did) and some things
worked, but many functions did not work properly as the Collection mode was
only intended for non-objects.