Fix stale version tracking in SlabAlloc leading to a use-after-free
SlabAlloc::m_youngest_live_version was only updated when committing a write
transaction via that SlabAlloc, which resulted in it having the incorrect value
if the write was made via a different DB instance. This lead to old ref
translations being marked with the wrong version, and then they'd be deleted
while they were still in use.