Avoid race when starting ExternalCommitHelper on windows
Previously if a notification came in between the time when the ECH was
contructed and when the background listener thread acquired the lock,
that notification would be missed. Now, the ECH looks at the
notification count on the constructing thread prior to launching the
background thread. If a notification comes in, the listener thread will
see a different notification count and will not wait on the condvar
prior to calling on_change().