Remove type coercion on bool and ObjectId when doing queries
There was also a request to remove type coercion betwwen string and
binary, but that would be more problematic. There has been a long tradition
for treating those two types as the same, so it would be a bit risky
to enforce separation. There is also little chance that it will produce
unexpected results in queries.