Ensure proper move operation is used for PayloadPolicy::Move
In ConstTableView constructor taking a const ConstTableView as src parameter
the copy assignment function was called in both the "Copy" and the "Move" case.
If the move assignment functions should be used, the src parameter should refer
to a writeable object as the content of m_key_values would be moved out of it.