Report if Realm::delete_files() actually deleted the Realm file
Our objective-c API for removing files returns true if the Realm file was
deleted, and sets an error out parameter if any errors occurred. This means
that even if an exception is thrown when trying to delete one of the auxiliary
files we still need to know if the main file was deleted, and that awkwardly
requires an out parameter.

This also refactors things to eliminate the unordered_map of auxiliary files
because none of the things using it actually needed a map, and consolidates the
two places where we deleted all the files into one.