Add collision handling in Dictionary
When a collision is detected, the new entry is created with a random key value with
the upper most bit set. This prevents it from colliding with future hashes.
These sibling entries are linked from the 'master' entry by a list in a
third column in the Dictionary.

When a master entry is deleted, the last sibling is copied over and becomes the
new master. When a sibling is deleted, it is just removed from the list in the
master.