Fix an assertion failure when an observed object is deleted
Because `TableKey::operator=(uint32_t)` exists and
`TableKey::operator=(TableKey)` doesn't, `m_table_key = {}` set the key to zero
instead of the null value, which resulted in `(bool)m_table_key` being true. As
a result, ObjectNotifier would continue to try to do things after the observed
object was deleted.

The simple fix for this is to remove `TableKey::operator=(uint32_t)` as it
doesn't appear to be used anyway.