Release Scheduler on Realm Deconstruction
In realm-js, we were witnessing a large delay when the process was killed.
It seems this was due to the cleanup being left to the garbage collector,
which would take anywhere between 5 to 10 seconds to release everything.

This change will release the resources immediately on realm closure and
fixes the issue we were having.