Workaround issue when initiating sync on a copied realm file (#4878)
The problem is that the server does not know that the client has already
integrated some server versions, and therefore it sends down 0 as
progress/upload/last_integrated_server_version. On the client this number
may be bigger and in this case that number will not be overwritten with 0.

On the server side, the following behavior is expected: If the IDENT message
indicates that the newly created client has already integraded some server
changesets, adjust the upload threshold accordingly- If the threshold is ahead
of the upload cursor stored in the history, use the threshold values.