Use the env TEMPDIR if available instead of P_tmpdir.
This does not affect Windows or Android.
On mac this changes the directory that test files are written from
"/var/tmp/" to "/var/folders/1y/_7..." which allows the OS to clean them
up on a reboot.

Additionally, `TestFile` will clean up the temp directory it creates
when finished, if not instructed to persist it.