Cut down on how many files indirectly import json.hpp
Due to that it defines the entire library in a header, json.hpp takes close to
a second to include even if nothing is used from it. A few headers included it
despite not using it, which resulted in 45 cpp files in total in core importing
it. These changes cut it to 10.