Reduce instantiations of std::sort and std::set
Each instantiation of std::sort() is about 3 KB, which isn't a ton, but Lst was
instantiating around 50 of them. We can reduce this to one at the cost of
introducing a single virtual call per comparison. Based on the single benchmark
of this, hoisting the load of `m_tree` out of the loop entirely cancels out the
perf hit from this.

There were also a handful of other places where we were pointlessly
instantiating std::sort() with different lambdas. ICF would make this not a
problem, but ld64 doesn't support ICF.