Fix user tokens not being threadsafe
This could manifest in a segfault if the access token refresh handler
was triggered at a specific time while the user is logged out from a
different thread. Specifically, resetting the refresh jwt at the same
time that it was being copied out by `SyncUser::refresh_jwt()` and
the util::Optional<BsonDocument> was read as being valid but it was just
set to util::none, and now the thread requesting the jwt triggers a copy
of invalid random data.