Fix some bugs in distinct collections
`Results::evaluate_sort_and_distinct_on_collection()` required rechecking if
`m_mode` was still `Collection` after calling it, and two places failed to do
so (`size()` and `clear()`), which meant that if those were the first functions
called on a newly constructed distinct collection of objects they'd actually
use the full collection.

This design was confusing and error-prone so I refactored the whole thing to
have a single `ensure_up_to_date()` function called at the start of each public
Results function which handles all of the mode-switching before the main switch
for the current operation.

There turned out to be another bug in `TableView::clear()` and
`TableView::remove()`. These functions marked the TableView as being in sync if
it was in sync at the start of the function, but this isn't actually true for
distinct TableViews. Removing an object from a distinct TableView can result in
a new object being added to replace it if there's more objects with the same
value in the distinct column, and so for distinct TableViews we need to call
do_sync() after remove()/clear().

I fixed this for both remove() and clear() but did not mention remove() in the
changelog because it appears to now be unused.