Fix a crash when calling LnkSet::remove_all_target_rows()
8f50047de890a55a799b34bcb0e4dfc097a02cec made it so that any functions in Set
which fail to call update() will crash when called on a newly-constructed Set
as m_tree is no longer initialized at construction time. This revealed that
LnkSet::remove_all_target_rows() was failing to do so (and not doing so was
always a problem; it just had a less obvious symptom previously).

This function was doubly untested: there weren't any tests which called this
function at all, and even if there had been they wouldn't have called it on a
newly-constructed instance and so wouldn't have caught the bug. I fixed both
parts of this by adding some more LnkSet tests and making it so that many of
the object store Set tests validate both that reusing a Set instance works and
that constructing a new Set instance every time works.