Remove virtual pointer-based state machine in SyncSession
This is mostly just code motion and simplification. The only behavior change
is that shutdown_and_wait() will now transition to Inactive from
WaitingForAccessToken. Previously in was a no-op in WaitingForAccessToken, but
I believe that was a bug that was just easy to miss with how the code used to
be written.