Spread query performance statistics over more than one cluster
The idea is that the index of the best node is stored in the QueryState
object along with the number of matches found by that node. When that
number reached 64, the other nodes will be probed and the distance
between hits will be calculated. Based on this the new best node will
be found.