Move low-level management of websocket connection out of ClientImpl::Connection
It is now moved to websocket::EZSocket which will eventually become the API
provided by the platform networking hooks. All of the EZ names are temporary
and will be replaced as part of providing the real platform networking API,
but I needed a clear way to separate the low-level websocket stuff from the
new APIs for now.

This is mostly just code motion and renaming. The main exception is that the
ConnectionTerminationReason enum had its resolve_operation_canceled and
connection_operation_canceled folded into closed_voluntarily, because they
were no longer observably different states, and they were always treated the
same anyway. Additionally the websocket message handlers now correctly return
false if they close the socket, to tell the websocket loop that it should
stop.