Stop trying to look at websocket PONG opcode messages
That was all dead code because the client currently does not send websocket
PING opcode messages. This is unrelated to the protocol-level "ping" and
"pong" messages that are embedded in the websocket BINARY opcode messages.

This gets us closer to being able to use stock websocket libs, because the
ping/pong opcode messages are considered a low-level implementation detail in
the protocol and are not necessarily exposed to code.