Fix TSAN lock-order-inversion warnings
Generally it is not needed to lock a mutex owned by an object in the
constructor of that object.

The main problem with lock-order inversion is that we have mutual
dependency between eg. SyncSession and SyncManager. If we call a
function on SyncSession that subsequently calls a function on SyncManager,
then we will first lock SyncSession::m_state_mutex and then SyncManager::m_mutex.
If we call a function on SyncManager that will call a function on SyncSession
the order will be the opposite. There are 3 solutions to this:

 - Try to unlock the mutex in the top-level function before calling further down
 - Use a separate mutexes in top-level and low-level functions
 - Avoid locking a mutex in the low-level function by using an atomic instead

Introduction of SyncSession::m_external_reference_mutex and SyncUser::m_tokens_mutex
is done in order to use separate mutexes.

SyncUser::m_state is turned into an atomic.