Add 'try_get_object' function to TableView
The original implementation in ObjList uses two virtual methods to
first find out if the key was valid and  - if it was - subsequently
find the object using that key. For tables, this can be done more
efficiently by just trying to look up the object.

At the same time TableView::get() is removed. It is silly to have two
names for the same function.