Filter out embedded orphan types earlier in the schema update process (#5115)
* Filter out embedded orphan types earlier in the schema update process

In AdditiveDiscovered mode we don't create tables for embedded object types which are never actually used. However, because of where we did this logic we would think that we had schema changes to apply every time we opened a Realm and so would commit an empty write transaction. https://github.com/realm/realm-core/pull/4965 changed this from a silent performance problem to a crash because starting the pointless write transaction would fail.

To fix this, filter out table additions for orphaned types in compare() rather than when applying the result of compare().

* const SubscriptionSet will be fixed separately later

Co-authored-by: James Stone <james.stone@mongodb.com>