Support sorting based on values from a dictionary.
Sorting is implemented by storing the parent dictionary's ColKey and the specific dictionary key we want to sort by.  This information is stored inside a new type of key called DictionaryKey.

A new variant-derived structure called SortableColumnKey was added to handle dispatching the key-specific behavior needed to support both ColKeys and DictionaryKeys in the Sorter. There is currently a workaround to a GCC bug, see https://stackoverflow.com/questions/51309467/using-stdvisit-on-a-class-inheriting-from-stdvariant-libstdc-vs-libc and https://gcc.gnu.org/bugzilla/show_bug.cgi?id=90943 for more information.

Sorting by direct dictionary keys and following links through a dictionary during sorting both work.

This is a work-in-progress, the performance characteristics of searching the dictionary for the desired key on every iteration of the sort predicate are unknown.

Sorting by dictionary keys using keyPath is currently not implemented (but desired).

In the future this could be extended to support sorting on specific indexes of a collection, ie. sort('array[0]'), by adding a new variant type to the SortableColumnKey structure and implementing the necessary visit lambdas.