Simplify UniqueFunction's SFINAE
The SFINAE workarounds for VS appear to have been for VS 2017 or older, as the
commit introducing them was from 2018. We currently target 2019, and aligning
the SFINAE check with what the standard library does for std::function in that
version improves compile performance a little and makes debugging less annoying.

Using make_unique<Derived> only to immediately store the result into a
unique_ptr<Base> results in unique_ptr<Derived> being instantiated. In debug
builds this nontrivially slows down the build.