Fix race in SubscriptionSet::get_state_change_nofication() (#5146)
If there have been writes to the database since SubscriptionSet was
created we need to refresh our state before maybe returning a ready
future. We also need to block process_notifications while determining
whether to return a ready future because otherwise the state may change
between deciding not to add a ready future and acquiring the lock to add
the pending future.