Support async writes with all schedulers and fix a use-after-free
Scheduler is not intended to be tightly coupled to Realm, and so should not
have several different notify functions for Realm-specific things. Instead,
redesign it to have a generic "invoke this function instead the scheduler"
function, and actually implement that in all of the Schedulers.

Some of the Scheduler callbacks captured a pointer to a Realm without any
guarantee that the Realm would still be alive when the callback was invoked.

The Android, dispatch queue, and C API schedulers did not support async writes.